requirements.txt
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.14
langchain-pinecone==0.2.5
openai==1.75.0
pinecone==6.0.2
python-dotenv==1.0.0
streamlit==1.29.0
tiktoken==0.9.0
pyyaml==6.0
schedule==1.2.0
argparse==1.4.0
pytest==7.3.1
pytest-mock==3.10.0 